fix(ui): update imports to renamed reactionWorkups.transform
Follow-up to the file rename: the converter and test now import from
./reactionWorkups.transform.ts (the rename commit landed without these edits).
